Buat transaksi laundry buat fungsi preview printer dan cetak printer

+++ app/src/main/java/dev/ferynnd/baguslaundry/model/ProductLaundry.kt
@@ -1,15 +1,17 @@
 package dev.ferynnd.baguslaundry.model
 
+import java.math.BigDecimal
+
 data class ProductLaundry(
     val id_laundry_item: Int? = null,
     val id_branch_laundry_item: Int? = null,
     val name_laundry_item: String?= null,
-    val price_laundry_item: Int? = null,
+    val price_laundry_item: BigDecimal? = null,
     val time_laundry_item: String? = null,
     val is_active_laundry_item: IsActiveLaundryItem? = IsActiveLaundryItem.active,
     val deleted_at: Any? = null,
     var isSelected: Boolean? = false,
-    var weight: Float? = null
+    var weight: BigDecimal? = null
 )
 
 enum class IsActiveLaundryItem {
